Small refactor of check_marathon_services_replication
Move one more function into monitoring_tools
