+++ paasta_tools/check_marathon_services_replication.py
@@ -36,7 +36,6 @@ from datetime import datetime
 from datetime import timedelta
 
 import a_sync
-import pysensu_yelp
 
 from paasta_tools import marathon_tools
 from paasta_tools import monitoring_tools
@@ -47,7 +46,6 @@ from paasta_tools.paasta_service_config_loader import PaastaServiceConfigLoader
 from paasta_tools.smartstack_tools import MesosSmartstackReplicationChecker
 from paasta_tools.utils import datetime_from_utc_to_local
 from paasta_tools.utils import DEFAULT_SOA_DIR
-from paasta_tools.utils import is_under_replicated
 from paasta_tools.utils import list_services
 from paasta_tools.utils import load_system_paasta_config
 
@@ -97,57 +95,10 @@ def check_healthy_marathon_tasks_for_service_instance(
         app_id=app_id,
     )
     log.info("Checking %s in marathon as it is not in smartstack" % app_id)
-    send_replication_event_if_under_replication(
+    monitoring_tools.send_replication_event_if_under_replication(
         instance_config=instance_config,
         expected_count=expected_count,
         num_available=num_healthy_tasks,
-    )
-
-
-def send_replication_event_if_under_replication(
-    instance_config,
-    expected_count,
-    num_available,
-):
-    crit_threshold = instance_config.get_replication_crit_percentage()
-    output = (
-        'Service %s has %d out of %d expected instances available!\n' +
-        '(threshold: %d%%)'
-    ) % (instance_config.job_id, num_available, expected_count, crit_threshold)
-    under_replicated, _ = is_under_replicated(num_available, expected_count, crit_threshold)
-    if under_replicated:
-        output += (
-            "\n\n"
-            "What this alert means:\n"
-            "\n"
-            "  This replication alert means that the service PaaSTA can't keep the\n"
-            "  requested number of copies up and healthy in the cluster.\n"
-            "\n"
-            "Reasons this might be happening:\n"
-            "\n"
-            "  The service may simply unhealthy. There also may not be enough resources\n"
-            "  in the cluster to support the requested instance count.\n"
-            "\n"
-            "Things you can do:\n"
-            "\n"
-            "  * Increase the instance count\n"
-            "  * Fix the cause of the unhealthy service. Try running:\n"
-            "\n"
-            "      paasta status -s %(service)s -i %(instance)s -c %(cluster)s -vv\n"
-        ) % {
-            'service': instance_config.service,
-            'instance': instance_config.instance,
-            'cluster': instance_config.cluster,
-        }
-        log.error(output)
-        status = pysensu_yelp.Status.CRITICAL
-    else:
-        log.info(output)
-        status = pysensu_yelp.Status.OK
-    monitoring_tools.send_replication_event(
-        instance_config=instance_config,
-        status=status,
-        output=output,
     )
